Support more than one digit versions in Makefile
The string comparisson of PG versions was returning true to any
comparisson of 10 or any other greater version with 9.x or 8.4 versions.
Trying to compare separately major and minor versions in order to be
able to support more than one digit versions.

Also, parse non released versions with alpha, beta or rc suffixes.
Ignore the suffix and keep the numeric part.